requirements.txt
# Data Processing
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0

# Visualization
plotly>=5.13.0
dash>=2.9.0
dash-bootstrap-components>=1.4.0

# Machine Learning
lightgbm>=3.3.0
shap>=0.41.0

# Data Storage
sqlalchemy>=1.4.0
pyyaml>=6.0

# Utilities
python-dotenv>=0.19.0
tqdm>=4.65.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Development
black>=22.0.0
flake8>=6.0.0
isort>=5.12.0 